Api calling using Retrofit done

+++ OnlineLearning/app/src/main/java/com/example/onlinelearning/model/ProfileData.kt
@@ -12,7 +12,7 @@ data class ProfileData(
         fun getProfileData(): ArrayList<ProfileData> {
             val profileData: ArrayList<ProfileData> = arrayListOf()
 
-            profileData.add(ProfileData(R.drawable.ic_dark_mode, "Dark Mode", Constants.ZERO.toInt()))
+            profileData.add(ProfileData(R.drawable.ic_dark_mode, "Dark Mode", Constants.ZERO))
             profileData.add(ProfileData(R.drawable.ic_settings, "Settings", Constants.ONE))
             profileData.add(ProfileData(R.drawable.ic_wallet, "Payment Details", Constants.ONE))
             profileData.add(ProfileData(R.drawable.ic_achievement, "Achievement", Constants.ONE))
